Split up BT page. Implemented CollectionView to Scanned devices list
